refactor(agent-engine): isolate tool execution runtime inputs
